Allow setting WiFI SSID as a command line argument
It might be needed to setup a different network for a specific
deployment and using a CLI argument is much faster than creating a new
configuration file, or modifying an existing one.

For example, Raspberry Pi Zero W requires a 2.4GHz only network and such
a network can have a different SSID than a regular multi-band one.